fix: in-memory repository condominium method delete and all spelling errors
