+++ src/modules/accounts/dtos/ICreateCondominiumDTO.ts
@@ -1,8 +1,8 @@
-interface ICreateCondominumDTO {
+interface ICreateCondominiumDTO {
   name: string;
   email: string;
   cnpj: string;
   firebase_id: string;
 }
 
-export { ICreateCondominumDTO };
+export { ICreateCondominiumDTO };
