--- src/modules/accounts/repositories/ICondominumRepository.ts
@@ -1,10 +1,0 @@
-import { ICreateCondominumDTO } from "../dtos/ICreateCondominumDTO";
-import { Condominiums } from "../infra/typeorm/entities/Condominiums";
-
-interface ICondominumRepository {
-  create(data: ICreateCondominumDTO): Promise<void>;
-  findAll(): Promise<Condominiums[]>;
-  delete(id: string): Promise<Condominiums>;
-}
-
-export { ICondominumRepository };
